Rename worker -> command_runner
